requirements.txt
streamlit
requests
msal
pdfplumber
python-docx